samples: openthread: Separate core CoAP Client functionality
Code refactoring which decouples dependencies and make the sample
be more clear.
