+++ samples/openthread/coap_client/Kconfig
@@ -11,3 +11,7 @@ endmenu
 module = COAP_CLIENT
 module-str = CoAP client
 source "${ZEPHYR_BASE}/subsys/logging/Kconfig.template.log_config"
+
+module = COAP_CLIENT_UTILS
+module-str = CoAP client utilities
+source "${ZEPHYR_BASE}/subsys/logging/Kconfig.template.log_config"
