+++ samples/openthread/coap_server/prj.conf
@@ -30,8 +30,6 @@ CONFIG_NET_PKT_RX_COUNT=16
 CONFIG_NET_PKT_TX_COUNT=16
 CONFIG_NET_BUF_RX_COUNT=80
 CONFIG_NET_BUF_TX_COUNT=80
-CONFIG_NET_CONTEXT_NET_PKT_POOL=y
-CONFIG_NET_MAX_CONTEXTS=10
 
 # Network shell
 CONFIG_NET_SHELL=y
