+++ samples/openthread/coap_server/sample.yaml
@@ -2,6 +2,6 @@ sample:
   description: Test CoAP communication in Thread network
   name: Openthread CoAP Server
 tests:
-  sample.net.openthread.coap_server:
+  sample.openthread.coap_server:
     tags: net openthread coap
     platform_whitelist: nrf52840dk_nrf52840 nrf52833dk_nrf52833
